Fix bad type passed to sscanf
--HG--
branch : develop

+++ code/ryzom/server/src/ai_service/sheets.cpp
@@ -581,7 +581,7 @@ void AISHEETS::CCreature::readGeorges(NLMISC::CSmartPtr<NLGEORGES::UForm> const&
 	}
 	if (item.getValueByName(s, "Basics.FameForGuardAttack") && !s.empty())
 	{
-		double tmp;
+		float tmp;
 		sscanf(s.c_str(), "%f", &tmp);
 		_FameForGuardAttack = (sint32)tmp;
 	}
